MeteorJS
Using Meteor.JS, i created a solution that populates a database using a
form through a webpage, this same webpage post's the top students in
each class but only when the refresh button is pressed, no postbacks
when the form is submited. In addition it provides an API for checking
the DataBase. It also received a small CSS to look presentable.
Instruction to instal Meteor.JS in windows also inside.